+++ benchmarks/runtime_refs/reduce_parallel/common/reduce_parallel_core.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void li_reduce_parallel_kernel(unsigned threads);
+double li_reduce_parallel_checksum(void);